Fixed converting of cs.

+++ crypto-scout-collector/src/main/java/com/github/akarazhev/cryptoscout/bybit/BybitServiceImpl.java
@@ -116,7 +116,7 @@ class BybitServiceImpl implements BybitService {
         }
         
         if (data.get(CS) != null) {
-            bybitTicker.setCs((Long) data.get(CS));
+            bybitTicker.setCs((Integer) data.get(CS));
         }
         // Process nested data object
         if (data.get(DATA) != null) {
